feat: refactor global menu structure in order to support unidirectal flow between modules
