+++ packages/frontend/src/components/PageLayout/GlobalMenu/shared.tsx
@@ -1,0 +1,31 @@
+import type { MenuItemProps } from '@altinn/altinn-components';
+import { Link } from 'react-router-dom';
+import { PageRoutes } from '../../../pages/routes.ts';
+
+export const isRouteSelected = (currentRoute: string, targetRoute: PageRoutes, fromView?: string) => {
+  if (currentRoute === targetRoute) {
+    return true;
+  }
+
+  if (fromView && targetRoute === fromView) {
+    return true;
+  }
+
+  /* default to inbox if no fromView and currentRoute is not a top level route, e.g. viewing a dialog entered the url */
+  if (
+    !fromView &&
+    !Object.values(PageRoutes).includes(currentRoute as PageRoutes) &&
+    targetRoute === PageRoutes.inbox
+  ) {
+    return true;
+  }
+
+  return false;
+};
+
+export const createMenuItemComponent =
+  ({ to, isExternal = false }: { to: string; isExternal?: boolean }): React.FC<MenuItemProps> =>
+  (props) => {
+    // @ts-ignore
+    return <Link {...props} to={to} {...(isExternal ? { target: '__blank', rel: 'noopener noreferrer' } : {})} />;
+  };
