+++ packages/frontend/src/featureFlags/FeatureFlags.ts
@@ -3,6 +3,8 @@ export enum FeatureFlagKeys {
   DisableBulkActions = 'DisableBulkActions',
   DisableFavoriteGroups = 'DisableFavoriteGroups',
   EnableProfilePages = 'EnableProfilePages',
+  EnableProfileLinkInGlobalMenu = 'EnableProfileLinkInGlobalMenu',
+  EnableAccessManagementLinkInGlobalMenu = 'EnableAccessManagementLinkInGlobalMenu',
 }
 
 export interface FeatureFlags {
@@ -10,6 +12,8 @@ export interface FeatureFlags {
   [FeatureFlagKeys.DisableBulkActions]: boolean;
   [FeatureFlagKeys.DisableFavoriteGroups]: boolean;
   [FeatureFlagKeys.EnableProfilePages]: boolean;
+  [FeatureFlagKeys.EnableProfileLinkInGlobalMenu]: boolean;
+  [FeatureFlagKeys.EnableAccessManagementLinkInGlobalMenu]: boolean;
 }
 
 export const featureFlags: FeatureFlags = {
@@ -17,4 +21,7 @@ export const featureFlags: FeatureFlags = {
   [FeatureFlagKeys.DisableBulkActions]: true,
   [FeatureFlagKeys.DisableFavoriteGroups]: true,
   [FeatureFlagKeys.EnableProfilePages]: !location.hostname.includes('af.altinn.no'),
+  [FeatureFlagKeys.EnableProfileLinkInGlobalMenu]: import.meta.env.DEV || location.hostname.includes('af.altinn.no'),
+  [FeatureFlagKeys.EnableAccessManagementLinkInGlobalMenu]:
+    import.meta.env.DEV || location.hostname.includes('af.altinn.no'),
 };
