+++ packages/frontend/src/i18n/resources/nb.json
@@ -42,6 +42,7 @@
   "altinn.beta.exit": "Ta meg tilbake",
   "altinn.beta.inbox": "Innboks beta",
   "altinn.beta.try": "Prøv ny innboks",
+  "altinn.access_management": "Tilgangsstyring",
   "betaModal.description": "Vi pusser opp innboksen, og du får sniktitte på en tidlig versjon, vi kaller det beta. Den er fortsatt under utvikling, så noen meldinger kan mangle. Vi håper likevel du vil prøve.",
   "betaModal.exitButton": "Gå til dagens innboks",
   "betaModal.fabAltText": "Om Beta",
